Grammatical/name fixes to pgBackRest Katacoda. (#80)
The main change is to consistently use pgBackRest when referring to the project rather than pgbackrest, PGBackRest, backrest, and pgBaseBackup.  This reduces confusion for the user and search engines.  pgbackrest should only be used when invoking the binary or specifically taking about the binary.

Also did a light edit for spelling and grammar.

My code editor automatically removes multiple final linefeeds so there is a bit of noise from that.  However, I think it would be better to remove them anyway.

I have not tested these changes on Katacoda so that should be done before committing.